Move TextDocumentHelper into separate file
